requirements.txt
asgiref==3.5.2
atomicwrites==1.4.1
attrs==22.1.0
colorama==0.4.5
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
pytz==2022.2.1
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.2
